+++ test/services/GuardianCapiTest.scala
@@ -1,14 +1,18 @@
 package services
 
-import java.time.LocalDate
+import java.time.{LocalDate, ZoneOffset}
 
+import com.gu.contentapi.client.model.v1.{Content, SearchResponse}
 import fixtures.TestEdition
 import model.editions.{CapiDateQueryParam, CapiPrefillQuery, Edition, PathType}
-import org.scalatest.{FunSuite, Matchers}
-import services.editions.{CollectionTemplatingHelper, EditionsTemplating}
+import org.scalatest.{BeforeAndAfterEach, FlatSpec, Matchers}
 import services.editions.prefills.{CapiPrefillTimeParams, CapiQueryTimeWindow, MetadataForLogging, PrefillParamsAdapter}
 
-class GuardianCapiTest extends FunSuite with Matchers {
+import scala.concurrent.{ExecutionContext, Future}
+
+class GuardianCapiTest extends FlatSpec with Matchers with BeforeAndAfterEach {
+
+  behavior of "geneneratePrefillQuery"
 
   private val issueDate = LocalDate.of(2019, 10, 5)
 
@@ -18,7 +22,7 @@ class GuardianCapiTest extends FunSuite with Matchers {
 
   private val contentPrefillTimeWindow: CapiQueryTimeWindow = timeWindowCfg.toCapiQueryTimeWindow(issueDate)
 
-  test("geneneratePrefillQuery") {
+  it should "prepare query to get ArticleItems for prefill on create issue request" in {
 
     val getPrefillParams = PrefillParamsAdapter(
       issueDate,
@@ -47,7 +51,9 @@ class GuardianCapiTest extends FunSuite with Matchers {
     actual shouldEqual expected
   }
 
-  test("prepareGetPrefillArticlesQuery") {
+  behavior of "prepareGetPrefillArticlesQuery"
+
+  it should "prepare query to get ArticleItems for prefill on Suggest Articles request" in {
 
     val currentPageCodes = List("123", "456")
 
@@ -76,6 +82,88 @@ class GuardianCapiTest extends FunSuite with Matchers {
       "&show-tags=all" +
       "&show-blocks=main" +
       "&from-date=2019-10-04T00%3A00%3A00Z"
+  }
+
+  behavior of "readAllSearchResponsePages"
+
+  implicit val ctx: ExecutionContext = ExecutionContext.global
+
+  private val ignoredStr = "ignore"
+  private val ignoredInt = -1
+  private val dummyContent = Content(
+    id = ignoredStr,
+    webTitle = ignoredStr,
+    webUrl = ignoredStr,
+    apiUrl = ignoredStr,
+  )
+  private val responseWith3TotalPagesWith2ItemsPerPage = SearchResponse(
+    status = ignoredStr,
+    userTier = ignoredStr,
+    total = 6,
+    startIndex = ignoredInt,
+    pageSize = 2,
+    currentPage = 1,
+    pages = 3,
+    orderBy = ignoredStr,
+    results = Seq(dummyContent, dummyContent)
+  )
+
+  private val responseWith1PageWith2ItemsPerPage = responseWith3TotalPagesWith2ItemsPerPage.copy(total = 2, pageSize = 2, pages = 1)
+  private val emptyResponse = responseWith3TotalPagesWith2ItemsPerPage.copy(total = 0, pageSize = 0, pages = 1, currentPage = 0, results = Nil)
+
+  private var getResponseCallsCounter = 0
+
+  override def beforeEach(): Unit = getResponseCallsCounter = 0
+
+  private val genFakeGetResponseFunction = (sr: SearchResponse) => (q: CapiQueryGenerator) => {
+    getResponseCallsCounter += 1
+    val curPage: Int = q.parameters.getOrElse("page", 1).toString.toInt
+    Future.successful(sr.copy(currentPage = curPage))
+  }
+
+
+  private val query = CapiQueryGenerator(PathType.Search).showTags("all")
+    .orderBy("newest")
+    .page(1)
+    .pageSize(2)
+    .tag("theguardian/mainsection/topstories")
+    .fromDate(LocalDate.of(2019, 11, 6).atStartOfDay().toInstant(ZoneOffset.UTC))
+    .toDate(LocalDate.of(2019, 11, 6).atStartOfDay().toInstant(ZoneOffset.UTC))
+    .useDate("newspaper-edition")
+
+  it should "read all pages of paginated response" in {
+    val getResponse = genFakeGetResponseFunction(responseWith3TotalPagesWith2ItemsPerPage)
+    val actual = GuardianCapi.readAllSearchResponsePages(query, getResponse)
+    actual.map(_.total) shouldEqual List(6, 6, 6)
+    actual.map(_.pageSize) shouldEqual List(2, 2, 2)
+    actual.map(_.currentPage) shouldEqual List(1, 2, 3)
+    actual.map(_.pages) shouldEqual List(3, 3, 3)
+    actual.flatMap(_.results).size shouldEqual 6
+    getResponseCallsCounter shouldEqual 3
+  }
+
+  it should "get response for 1 page search result" in {
+    val getResponse = genFakeGetResponseFunction(responseWith1PageWith2ItemsPerPage)
+    val actual = GuardianCapi.readAllSearchResponsePages(query, getResponse)
+    actual.map(_.total) shouldEqual List(2)
+    actual.map(_.pageSize) shouldEqual List(2)
+    actual.map(_.currentPage) shouldEqual List(1)
+    actual.map(_.pages) shouldEqual List(1)
+    actual.flatMap(_.results).size shouldEqual 2
+    getResponseCallsCounter shouldEqual 1
+  }
+
+  it should "handle situation if search result pages is 0" in {
+    val getResponse = genFakeGetResponseFunction(emptyResponse)
+    val actual = GuardianCapi.readAllSearchResponsePages(query, getResponse)
+
+    actual.map(_.total) shouldEqual List(0)
+    actual.map(_.pageSize) shouldEqual List(0)
+    actual.map(_.currentPage) shouldEqual List(1)
+    actual.map(_.pages) shouldEqual List(1)
+    actual.flatMap(_.results).size shouldEqual 0
+
+    getResponseCallsCounter shouldEqual 1
   }
 
 }
